# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/cases/{caseId}/start-preparation": {
      "post": {
        "operationId": "CasesController_startPreparation",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Begin handling a package (assigned → in_progress, case.started) or resume after Teamlead clearance (problem_resolved → in_progress, case.resumed)",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{caseId}/complete": {
      "post": {
        "operationId": "CasesController_complete",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Complete a package (in_progress → completed, case.completed)",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{caseId}/partial-complete": {
      "post": {
        "operationId": "CasesController_partialComplete",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartialCompleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Teilabschluss mit gesammelten Problemen (in_progress → issue_open, case.problems_reported); der Beleg bleibt beim selben MA geparkt",
        "tags": [
          "cases"
        ]
      }
    },
...
    "/api/teamlead/cases": {
      "get": {
        "operationId": "TeamleadController_pool",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by CaseStatus",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "section",
            "required": false,
            "in": "query",
            "description": "Filter by section code",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Volltext: WE-Beleg-Nr / Lagerplatz-Code / Lieferschein-Nr (contains)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shopNo",
            "required": false,
            "in": "query",
            "description": "Filter: Shop (primärer Shop, contains)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchNo",
            "required": false,
            "in": "query",
            "description": "Filter: Filiale (contains)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bereich",
            "required": false,
            "in": "query",
            "description": "Filter: fester Bereich (Hängebahn|Palette|Regal)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assigned",
            "required": false,
            "in": "query",
            "description": "Filter: zugeteilt ja/nein",
            "schema": {
              "type": "string",
              "enum": [
                "yes",
                "no"
              ]
            }
          },
          {
            "name": "labels",
            "required": false,
            "in": "query",
            "description": "Filter: Etiketten nötig ja/nein",
            "schema": {
              "type": "string",
              "enum": [
                "yes",
                "no"
              ]
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "description": "Lebenszyklus-Scope (server-seitig): aktiv | abgeschlossen | archiv (completed+zst_done) | topf (Aufmerksamkeit/blocked/needs_review) | alle",
            "schema": {
              "type": "string",
              "enum": [
                "aktiv",
                "abgeschlossen",
                "archiv",
                "topf",
                "alle"
              ]
            }
          },
          {
            "name": "bookingFrom",
            "required": false,
            "in": "query",
            "description": "Buchungsdatum ab (YYYY-MM-DD, inklusive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookingTo",
            "required": false,
            "in": "query",
            "description": "Buchungsdatum bis (YYYY-MM-DD, inklusive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Server-side sort column",
            "schema": {
              "type": "string",
              "enum": [
                "weBelegNo",
                "bookingDate",
                "totalQuantity",
                "effortPoints",
                "status",
                "section",
                "branchNo",
                "primaryShopNo",
                "completedAt"
              ]
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoolListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the operational pool (filter + paginate)",
        "tags": [
          "teamlead"
        ]
      }
    },
...
    "/api/teamlead/cases/{caseId}": {
      "get": {
        "operationId": "TeamleadController_caseDetail",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "§10.4 Belegdetails: one case with positions, boxes, documents, history",
        "tags": [
          "teamlead"
        ]
      }
    },
...
  },
  "info": {
    "title": "Paketlagerdispo Backend API",
    "description": "Digitale Belegverteilung – modular monolith backbone. Status state machine (§7), RBAC (§5/§16.1), audit event log (§7.2), and the case pool/lifecycle (§14.2).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "OIDC access token"
      }
    },
    "schemas": {
...
      "CaseSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "weBelegNo": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "Anhang A CaseStatus"
          },
          "section": {
            "type": "number",
            "nullable": true,
            "description": "Section 1|2|3|4|7|8, null for prio-only"
          },
          "priorityFlags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totalQuantity": {
            "type": "number"
          },
          "estimatedMinutes": {
            "type": "number"
          },
          "storageLocationCode": {
            "type": "string",
            "nullable": true,
            "description": "null nur bei blocked-Belegen (Intake-Gate D1: Lagerplatz fehlt)"
          },
          "storageLocationKind": {
            "type": "string",
            "nullable": true,
            "description": "Lagerklasse (LocationKind: regal|palette_*|haengebahn|…) — Quelle der Bereich-Icons"
          },
          "priceLabelPrintRequired": {
            "type": "boolean",
            "nullable": true,
            "description": "Preisetiketten müssen gedruckt werden (Arbeitsanweisung) — Hinweis beim Ware holen"
          },
          "primaryShopNo": {
            "type": "string",
            "nullable": true,
            "description": "Primärer Shop (A7)"
          },
          "inboundCartonCount": {
            "type": "number",
            "nullable": true,
            "description": "Kartons der Anlieferung (A6)"
          },
          "missingFields": {
            "description": "Intake-Gate: fehlende Pflichtfelder (blocked)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bookingDate": {
            "type": "string",
            "description": "ISO date YYYY-MM-DD"
          },
          "goodsType": {
            "type": "string",
            "nullable": true,
            "description": "GoodsTypeText (Warenart), null if unknown"
          },
          "assignedEmployeeName": {
            "type": "string",
            "nullable": true,
            "description": "Display name of the assigned employee"
          },
          "branchNo": {
            "type": "string",
            "description": "Filiale (Beleg-Kopf)"
          },
          "labelsRequired": {
            "type": "boolean",
            "description": "Etiketten nötig (abgeleitet: workInstruction.priceLabelPrintRequired || boxLabelRequired)"
          },
          "shopNos": {
            "description": "Alle Shops des Belegs (distinct über die Positionen, Primär-Shop zuerst)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "docuWareUrl": {
            "type": "string",
            "nullable": true,
            "description": "Link ins DocuWare-Langzeitarchiv (A6, mock); gesetzt bei Abschluss"
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "description": "ISO-8601 Abschlusszeitpunkt (completed/zst_done), sonst null"
          },
          "attentionFlag": {
            "type": "boolean",
            "description": "TL-Topf (A7): „Besondere Aufmerksamkeit\" (Bucherinnen-Inlet)"
          },
          "attentionNote": {
            "type": "string",
            "nullable": true,
            "description": "Notiz zum Aufmerksamkeitsflag"
          },
          "forwardedTo": {
            "type": "string",
            "nullable": true,
            "description": "Digitale Ablage (C5): Weiterleitungs-Empfänger (retourenabteilung|lieferscheinbucher); null = nicht weitergeleitet"
          }
        },
        "required": [
          "id",
          "weBelegNo",
          "status",
          "priorityFlags",
          "totalQuantity",
          "estimatedMinutes",
          "missingFields",
          "bookingDate",
          "branchNo",
          "labelsRequired",
          "shopNos",
          "attentionFlag"
        ]
      },
...
      "WorkInstructionHeaderDto": {
        "type": "object",
        "properties": {
          "priceLabelPrintRequired": {
            "type": "boolean"
          },
          "sortByArticleColorSizeRequired": {
            "type": "boolean"
          },
          "goodsReceiptCheckMode": {
            "type": "string",
            "description": "CheckMode: quantity_only|percentage_check|full_check"
          },
          "goodsReceiptCheckPercentage": {
            "type": "number",
            "nullable": true
          },
          "inspectionLevelCode": {
            "type": "string",
            "nullable": true,
            "description": "Prüfstufe: none|p10|p20|full (A5)"
          },
          "inspectionLevelLabel": {
            "type": "string",
            "nullable": true,
            "description": "Prüfstufen-Label, z. B. \"20 %\""
          },
          "inspectionDescription": {
            "type": "string",
            "nullable": true,
            "description": "Aufgabentext der Prüfstufe (welche Todos sie bedeutet)"
          },
          "minimumQuantityCheckAlwaysRequired": {
            "type": "boolean"
          },
          "boxLabelRequired": {
            "type": "boolean"
          },
          "zstRequired": {
            "type": "boolean"
          }
        },
        "required": [
          "priceLabelPrintRequired",
          "sortByArticleColorSizeRequired",
          "goodsReceiptCheckMode",
          "minimumQuantityCheckAlwaysRequired",
          "boxLabelRequired",
          "zstRequired"
        ]
      },
...
      "TransportBoxTargetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "boxNo": {
            "type": "number"
          },
          "branchNo": {
            "type": "string"
          },
          "shopAreaNo": {
            "type": "string"
          },
          "shopNo": {
            "type": "string",
            "nullable": true
          },
          "floor": {
            "type": "string",
            "nullable": true
          },
          "goodsType": {
            "type": "string",
            "nullable": true,
            "description": "BoxGoodsType"
          },
          "positionIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "plannedQuantity": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "labelStatus": {
            "type": "string",
            "description": "BoxLabelStatus: not_required|pending|printed|reprinted"
          },
          "sealed": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "boxNo",
          "branchNo",
          "shopAreaNo",
          "positionIds",
          "plannedQuantity",
          "quantity",
          "labelStatus",
          "sealed"
        ]
      },
...
      "TransitionResultDto": {
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "eventId": {
            "type": "object",
            "nullable": true,
            "description": "Audit event id, if a milestone was recorded"
          }
        },
        "required": [
          "caseId",
          "status",
          "version"
        ]
      },
      "SkuQuantityDto": {
        "type": "object",
        "properties": {
          "skuLineId": {
            "type": "string"
          },
          "confirmedQuantity": {
            "type": "number",
            "description": "Gezählte Ist-Menge der Größenzeile"
          },
          "correctedVkPrice": {
            "type": "number",
            "description": "Korrigierter VK, wenn der Etikettpreis falsch ist (Preisabweichung ⇒ implizites Problem)"
          }
        },
        "required": [
          "skuLineId",
          "confirmedQuantity"
        ]
      },
      "CompleteDto": {
        "type": "object",
        "properties": {
          "skuQuantities": {
            "description": "Gezählte Ist-Mengen; ohne Angabe gilt Ist=Soll",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkuQuantityDto"
            }
          }
        }
      },
      "ReportedProblemDto": {
        "type": "object",
        "properties": {
          "positionId": {
            "type": "string",
            "description": "ReceiptPosition, auf die sich das Problem bezieht"
          },
          "skuLineId": {
            "type": "string",
            "description": "Optional auf eine Größenzeile eingegrenzt"
          },
          "reasonId": {
            "type": "string",
            "description": "ProblemReason-Katalog-Eintrag (aktiv)"
          },
          "note": {
            "type": "string",
            "description": "Freitext-Notiz des MA zum Problem"
          }
        },
        "required": [
          "positionId",
          "reasonId"
        ]
      },
      "PartialCompleteDto": {
        "type": "object",
        "properties": {
          "skuQuantities": {
            "description": "Gezählte Ist-Mengen aller Größenzeilen",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkuQuantityDto"
            }
          },
          "problems": {
            "description": "Manuell erfasste Positions-Probleme",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportedProblemDto"
            }
          }
        },
        "required": [
          "skuQuantities",
          "problems"
        ]
      },
...
      "PoolItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "weBelegNo": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "Anhang A CaseStatus"
          },
          "section": {
            "type": "number",
            "nullable": true,
            "description": "Section 1|2|3|4|7|8, null for prio-only"
          },
          "priorityFlags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totalQuantity": {
            "type": "number"
          },
          "estimatedMinutes": {
            "type": "number"
          },
          "storageLocationCode": {
            "type": "string",
            "nullable": true,
            "description": "null nur bei blocked-Belegen (Intake-Gate D1: Lagerplatz fehlt)"
          },
          "storageLocationKind": {
            "type": "string",
            "nullable": true,
            "description": "Lagerklasse (LocationKind: regal|palette_*|haengebahn|…) — Quelle der Bereich-Icons"
          },
          "priceLabelPrintRequired": {
            "type": "boolean",
            "nullable": true,
            "description": "Preisetiketten müssen gedruckt werden (Arbeitsanweisung) — Hinweis beim Ware holen"
          },
          "primaryShopNo": {
            "type": "string",
            "nullable": true,
            "description": "Primärer Shop (A7)"
          },
          "inboundCartonCount": {
            "type": "number",
            "nullable": true,
            "description": "Kartons der Anlieferung (A6)"
          },
          "missingFields": {
            "description": "Intake-Gate: fehlende Pflichtfelder (blocked)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bookingDate": {
            "type": "string",
            "description": "ISO date YYYY-MM-DD"
          },
          "goodsType": {
            "type": "string",
            "nullable": true,
            "description": "GoodsTypeText (Warenart), null if unknown"
          },
          "assignedEmployeeName": {
            "type": "string",
            "nullable": true,
            "description": "Display name of the assigned employee"
          },
          "branchNo": {
            "type": "string",
            "description": "Filiale (Beleg-Kopf)"
          },
          "labelsRequired": {
            "type": "boolean",
            "description": "Etiketten nötig (abgeleitet: workInstruction.priceLabelPrintRequired || boxLabelRequired)"
          },
          "shopNos": {
            "description": "Alle Shops des Belegs (distinct über die Positionen, Primär-Shop zuerst)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "docuWareUrl": {
            "type": "string",
            "nullable": true,
            "description": "Link ins DocuWare-Langzeitarchiv (A6, mock); gesetzt bei Abschluss"
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "description": "ISO-8601 Abschlusszeitpunkt (completed/zst_done), sonst null"
          },
          "attentionFlag": {
            "type": "boolean",
            "description": "TL-Topf (A7): „Besondere Aufmerksamkeit\" (Bucherinnen-Inlet)"
          },
          "attentionNote": {
            "type": "string",
            "nullable": true,
            "description": "Notiz zum Aufmerksamkeitsflag"
          },
          "forwardedTo": {
            "type": "string",
            "nullable": true,
            "description": "Digitale Ablage (C5): Weiterleitungs-Empfänger (retourenabteilung|lieferscheinbucher); null = nicht weitergeleitet"
          },
          "assignedEmployeeNo": {
            "type": "object",
            "nullable": true
          },
          "effortPoints": {
            "type": "number"
          },
          "deliveryGroup": {
            "nullable": true,
            "description": "Delivery-group context so groups are visible BEFORE distribution; null if standalone",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DeliveryGroupRefDto"
              }
            ]
          },
          "bereich": {
            "type": "string",
            "nullable": true,
            "description": "Beleg's fixed Bereich (Hängebahn|Palette|Regal), derived from the Lagerplatz kind; null for non-pickup kinds."
          },
          "bundleQueue": {
            "nullable": true,
            "description": "A5: Position des Belegs in seinem Bündel („vorbereitet · Pos n\"); null wenn nicht gebündelt",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/BundleQueueRefDto"
              }
            ]
          },
          "openIssue": {
            "nullable": true,
            "description": "C4: neuestes OFFENES Problem (Art + Notiz-Vorschau); null ohne offenes Problem",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OpenIssueRefDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "weBelegNo",
          "status",
          "priorityFlags",
          "totalQuantity",
          "estimatedMinutes",
          "missingFields",
          "bookingDate",
          "branchNo",
          "labelsRequired",
          "shopNos",
          "attentionFlag",
          "effortPoints"
        ]
      },
      "PoolListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PoolItemDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ]
      },
...
      "EffortComponentsDto": {
        "type": "object",
        "properties": {
          "base": {
            "type": "number",
            "description": "Grundzeit je Beleg"
          },
          "quantity": {
            "type": "number",
            "description": "Mengenerfassung"
          },
          "priceLabelPrint": {
            "type": "number",
            "description": "Etiketten drucken"
          },
          "labelAttach": {
            "type": "number",
            "description": "Etiketten anbringen"
          },
          "security": {
            "type": "number",
            "description": "Warensicherung"
          },
          "online": {
            "type": "number",
            "description": "Online-Behandlung"
          },
          "redPrice": {
            "type": "number",
            "description": "Rotpreis-Auszeichnung"
          },
          "check": {
            "type": "number",
            "description": "Prüfung (Mehraufwand)"
          },
          "handling": {
            "type": "number",
            "description": "Handling / Füllmaterial"
          }
        },
        "required": [
          "base",
          "quantity",
          "priceLabelPrint",
          "labelAttach",
          "security",
          "online",
          "redPrice",
          "check",
          "handling"
        ]
      },
      "PositionDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "positionNo": {
            "type": "number"
          },
          "wgr": {
            "type": "string",
            "description": "Warengruppe"
          },
          "wgrDescription": {
            "type": "string",
            "nullable": true,
            "description": "WGR-Klartext, z. B. \"D-Bermuda\" (A2)"
          },
          "catMan": {
            "type": "boolean",
            "nullable": true,
            "description": "CatMan-Kennzeichen (Anzeige, A3)"
          },
          "supplierColor": {
            "type": "string"
          },
          "expectedQuantity": {
            "type": "number",
            "description": "Σ expected over the position SKU lines"
          },
          "confirmedQuantity": {
            "type": "number",
            "nullable": true,
            "description": "Σ confirmed over the SKU lines, null if none confirmed yet"
          },
          "priceLabelRequired": {
            "type": "boolean"
          },
          "securityRequired": {
            "type": "boolean"
          },
          "onlineHandlingRequired": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "description": "PositionStatus: open|confirmed|issue_open|completed"
          },
          "skuLines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkuLineDto"
            }
          }
        },
        "required": [
          "id",
          "positionNo",
          "wgr",
          "supplierColor",
          "expectedQuantity",
          "priceLabelRequired",
          "securityRequired",
          "onlineHandlingRequired",
          "status",
          "skuLines"
        ]
      },
...
      "ZstSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "completedQuantity": {
            "type": "number",
            "description": "Confirmed quantity booked by this ZST"
          },
          "effortPoints": {
            "type": "number"
          },
          "completedAt": {
            "type": "string",
            "description": "ISO-8601 timestamp the ZST was set"
          },
          "exportedAt": {
            "type": "string",
            "nullable": true,
            "description": "ISO-8601 timestamp the ZST batch was exported (zst_done), else null"
          },
          "source": {
            "type": "string",
            "description": "ZstSource: mobile_app|teamlead_dashboard|manual_import"
          }
        },
        "required": [
          "id",
          "completedQuantity",
          "effortPoints",
          "completedAt",
          "source"
        ]
      },
...
      "CaseDetailDto": {
        "type": "object",
        "properties": {
          "case": {
            "$ref": "#/components/schemas/CaseSummaryDto"
          },
          "effortPoints": {
            "type": "number",
            "description": "Effort points (Aufwandspunkte)"
          },
          "effortComputed": {
            "type": "boolean",
            "description": "true = Aufwand live aus der Arbeitsanweisung berechnet; false = gespeicherter Schätzwert"
          },
          "effortComponents": {
            "nullable": true,
            "description": "Per-Treiber-Minutenaufschlüsselung; null beim gespeicherten Schätzwert",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/EffortComponentsDto"
              }
            ]
          },
          "deliveryNoteNo": {
            "type": "string",
            "nullable": true
          },
          "primaryShopAreaNo": {
            "type": "string",
            "nullable": true
          },
          "primaryFloor": {
            "type": "string",
            "nullable": true
          },
          "catManDate": {
            "type": "string",
            "nullable": true,
            "description": "ISO date YYYY-MM-DD"
          },
          "loadPlanDate": {
            "type": "string",
            "nullable": true,
            "description": "ISO date YYYY-MM-DD"
          },
          "goodsType": {
            "type": "string",
            "nullable": true,
            "description": "GoodsTypeText (Warenart)"
          },
          "workInstruction": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/WorkInstructionHeaderDto"
              }
            ]
          },
          "positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PositionDetailDto"
            }
          },
          "transportBoxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransportBoxTargetDto"
            }
          },
          "issues": {
            "description": "Reported problems, newest first",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueSummaryDto"
            }
          },
          "zstRecords": {
            "description": "ZST completion records, oldest first",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ZstSummaryDto"
            }
          },
          "history": {
            "description": "Audit history, newest first",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditEventDto"
            }
          },
          "deliveryGroup": {
            "nullable": true,
            "description": "Zugehörige Lieferung (Teamlead-Punkt 1): siblings + who holds them; null if standalone",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DeliveryGroupDetailDto"
              }
            ]
          }
        },
        "required": [
          "case",
          "effortPoints",
          "effortComputed",
          "positions",
          "transportBoxes",
          "issues",
          "zstRecords",
          "history"
        ]
      },
...
    }
  }
}